trie template argument compiles.  Untested.
git-svn-id: file:///dev/shm/somefilter.svn@407 e102df66-1e2e-11dd-9b44-c24451a4db5e

+++ lm/ngram.hh
@@ -56,7 +56,7 @@ template <class Search, class VocabularyT> class GenericModel : public base::Mod
     // Get the size of memory that will be mapped given ngram counts.  This
     // does not include small non-mapped control structures, such as this class
     // itself.  
-    static size_t Size(const std::vector<size_t> &counts, const Config &config = Config());
+    static size_t Size(const std::vector<uint64_t> &counts, const Config &config = Config());
 
     GenericModel(const char *file, const Config &config = Config());
